changes in services page
moved the jsx file to a new folder and made seperate components inside of it
